Inline useTemplates
